+++ recipes/lora_finetune_distributed.py
@@ -33,7 +33,6 @@ from torchtune.modules.peft.peft_utils import (
 )
 from torchtune.recipe_interfaces import FTRecipeInterface
 
-from torchtune.utils.distributed import validate_no_params_on_meta_device
 from tqdm import tqdm
 
 log = utils.get_logger("DEBUG")
@@ -325,7 +324,7 @@ class LoRAFinetuneRecipeDistributed(FTRecipeInterface):
         )
 
         # Ensure no params and buffers are on meta device
-        validate_no_params_on_meta_device(model)
+        utils.validate_no_params_on_meta_device(model)
 
         if enable_activation_checkpointing:
             utils.set_activation_checkpointing(
