+++ torchtune/utils/checkpoint.py
@@ -19,7 +19,7 @@ from torchtune.utils.constants import (
     SEED_KEY,
     TOTAL_EPOCHS_KEY,
 )
-from torchtune.utils.distributed import get_world_size_and_rank
+from torchtune.utils._distributed import get_world_size_and_rank
 
 
 def _contains_fsdp(model: nn.Module) -> bool:
